refac(fe):엔터 사용시 다음 순서에 텍스트 블록 추가 로직 구현
