+++ frontend/nevvel/src/store/EditorTextStore.tsx
@@ -1,0 +1,4 @@
+import { atom } from "jotai";
+import { episode,content } from "editor";
+
+export const blockTextAtom = atom<string>("")
